# my-concierge-server/static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "The MyConcierge API powers the MyConcierge web application by enabling the application to understand user desires and retrieve appropriate restaurant recommendations.",
        "version": "1.0.0",
        "title": "MyConcierge API"
    },
...
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/prompt": {
            "get": {
                "tags": [
                    "NLP API"
                ],
                "summary": "Analyze User Queries",
                "description": "Uses OpenAI's GPT to analyze user queries, updating the user preference vector with this information and determining the appropriate response for the chatbot to give to the user.",
                "parameters": [
                    {
                        "name": "prompt",
                        "in": "query",
                        "description": "User Query to be Analyzed",
                        "required": true,
                        "type": "string",
                        "example": "I would like some American food."
                    },
                    {
                        "name": "user_preference_vector",
                        "in": "query",
                        "description": "Underscore-Separated User Preference Vector",
                        "required": true,
                        "type": "string",
                        "example": "0_0_0_0_0_0_0_0_0_0_0_0_0_0_0_0_0_0_0_0_0_0_0_0_0_0_0_0_0_0_0_0_0_0_0"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "type": "string",
                                            "description": "Chatbot Response to User Query",
                                            "example": "Classic American fare! You can't go wrong with burgers and fries. What else can I assist you with?"
                                        },
                                        "updated_user_preference_vector": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer"
                                            },
                                            "description": "Underscore-Separated User Preference Vector Updated Based on User Query",
                                            "example": [5, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND"
                    }
                }
            }
        },
...
    }
}